chore: move buyers page to www
